refactor : roomservice의 game 메서드 gameservice로 이동
